start/requirements.txt
--find-links https://download.pytorch.org/whl/torch_stable.html

azure-functions
torch==1.9.0+cpu
torchvision==0.10.0+cpu
requests
pillow
onnxruntime
pandas
scipy
opencv-python
pyyaml
tqdm
matplotlib
seaborn
